backend/requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
-e git+https://github.com/shubham151/iskul@913a76b89b7a99154612d4c3ede8d6a27cd16373#egg=backend&subdirectory=backend
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
Django==5.0.4
dm-tree==0.1.8
executing==2.0.1
faiss-cpu==1.8.0
Farama-Notifications==0.0.4
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
groq==0.9.0
grpcio==1.62.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
image==1.5.33
ipykernel==6.29.3
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
lupyne==3.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
nest-asyncio==1.6.0
networkx==3.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.3.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.0
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.35.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
streamlit==1.33.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
tensorflow-intel==2.16.0rc0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2024.4.24
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.42.0
whisper==1.1.10
wrapt==1.16.0
youtube-dl==2021.12.17
